add back plaid add

+++ pages/api/remove_access_token.tsx
@@ -8,11 +8,21 @@ export default async (req, res) => {
     const response = await plaidClient.itemRemove({
       access_token: access_token
     });
-    await prisma.plaid.delete({
-      where: {
-        access_token: access_token,
-      },
-    })
+    if(response.data){
+      await prisma.plaid.delete({
+        where: {
+          access_token: access_token,
+        },
+      })
+      await prisma.accounts.updateMany({
+        where: {
+          access_token: access_token
+        },
+        data: {
+          active: false
+        }
+      })
+    }
     return res.status(200).json(response.data);
   } catch (error) {
     return res.status(500).json({ error: error.message || error.toString() })
